feat: implement os file directory and automate the csv dataset processing
